backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# LangChain v0.3
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-ollama==0.2.1
langchain-chroma==0.1.4

# Vector Store
chromadb==0.5.20

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
alembic==1.14.0

# Document Parsing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
xlrd==2.0.1
python-pptx==1.0.2
pillow==11.0.0
paddleocr==2.9.1
paddlepaddle==2.6.2

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
httpx==0.28.1
loguru==0.7.3
python-dotenv==1.0.1
chardet==5.2.0
